+++ templates/Uwp/Pages/Camera.CaliburnMicro/Views/CameraViewPage.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Navigation;
 
 namespace Param_ItemNamespace.Views
 {
