+++ templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_SplitView.TabbedPivot/.template.config/template.json
@@ -3,14 +3,14 @@
   "classifications": [
     "Universal"
   ],
-  "name": "CaliburnMicro.DeepLinking.TabbedPivot",
+  "name": "CaliburnMicro.DeepLinking.SplitView.TabbedPivot",
   "tags": {
     "language": "C#",
     "type": "item",
     "wts.type": "composition",
     "wts.platform" : "Uwp",
     "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & identity == wts.Feat.DeepLinking & $projectType == TabbedPivot"
+    "wts.compositionFilter": "$framework == CaliburnMicro & identity == wts.Feat.DeepLinking & $projectType == SplitView|TabbedPivot"
   },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
