+++ templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_SplitView.TabbedPivot/Strings/en-us/Resources_postaction.resw
