+++ templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_SplitView.TabbedPivot/Views/SchemeActivationSamplePage_postaction.xaml.cs
