--- templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_TabbedPivot/Helpers/IPivotActivationPage.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Param_ItemNamespace.Helpers
-{
-    public interface IPivotActivationPage
-    {
-        Task OnPivotActivatedAsync(Dictionary<string, string> parameters);
-    }
-}
