--- templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_TabbedPivot/Views/PivotPage_postaction.xaml.cs
@@ -1,27 +1,0 @@
-﻿//{[{
-using Param_ItemNamespace.Activation;
-//}]}
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class PivotPage : Page
-    {
-//{[{
-        private PivotViewModel ViewModel
-        {
-            get { return DataContext as PivotViewModel; }
-        }
-//}]}
-
-        protected override async void OnNavigatedTo(NavigationEventArgs e)
-        {
-            base.OnNavigatedTo(e);
-//{[{
-            if (e.Parameter is SchemeActivationData activationData)
-            {
-                ViewModel.ActivationData = activationData;
-            }
-
-//}]}
-        }
-    }
-}
