--- templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_TabbedPivot/Views/SchemeActivationSamplePage_searchreplace.xaml.cs
@@ -1,3 +1,0 @@
-﻿public sealed partial class SchemeActivationSamplePage : Page
-^^^-searchabove-replacebelow-vvv
-public sealed partial class SchemeActivationSamplePage : Page, IPivotActivationPage
