+++ templates/Uwp/_composition/CaliburnMicro/Feature.DeepLinking_TabbedPivot/Views/ShellPage_postaction.xaml
@@ -1,0 +1,18 @@
+﻿<!--{**-->
+<!--This code adds SchemeActivationSamplePage to the NavigationView control in ShellPage -->
+<!--**}-->
+
+<Page
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <winui:NavigationView
+        Background="{ThemeResource SystemControlBackgroundAltHighBrush}">
+        <winui:NavigationView.MenuItems>
+            <!--^^-->
+            <!--{[{-->
+            <winui:NavigationViewItem x:Uid="Shell_SchemeActivationSample" helpers:NavHelper.NavigateTo="views:SchemeActivationSamplePage" />
+            <!--}]}-->
+        </winui:NavigationView.MenuItems>        
+    </winui:NavigationView>
+</Page>
