--- templates/Uwp/_composition/CaliburnMicro/Page.TabbedPivot.AddBehavior/Helpers/IPivotPage.cs
@@ -1,11 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Param_ItemNamespace.Helpers
-{
-    public interface IPivotPage
-    {
-        Task OnPivotSelectedAsync();
-
-        Task OnPivotUnselectedAsync();
-    }
-}
