--- templates/Uwp/_composition/CaliburnMicro/Page.TabbedPivot.AddBehavior/Views/PivotPage_postaction.xaml
@@ -1,20 +1,0 @@
-﻿<Page
-    Style="{StaticResource PageStyle}"
-    <!--{[{-->
-    xmlns:behaviors="using:Param_ItemNamespace.Behaviors"
-    xmlns:i="using:Microsoft.Xaml.Interactivity"
-    <!--}]}-->
-    xmlns:views="using:Param_ItemNamespace.Views"
-    mc:Ignorable="d">
-
-    <Grid>
-
-        <Pivot x:Uid="PivotPage" x:Name="Items">
-            <!--{[{-->
-            <i:Interaction.Behaviors>
-                <behaviors:PivotBehavior />
-            </i:Interaction.Behaviors>
-            <!--}]}-->
-        </Pivot>
-    </Grid>
-</Page>
