+++ templates/Uwp/_composition/CaliburnMicro/Project.Blank.AddSystemBackButton/.template.config/template.json
@@ -1,21 +1,19 @@
 ﻿{
-  "author": "Microsoft Community",
+  "author": "Microsoft",
   "classifications": [
     "Universal"
   ],
-  "name": "CaliburnMicro.DeepLinking.TabbedPivot",
+  "name": "CaliburnMicro.Project.Blank.AddSystemBackButton",
   "tags": {
     "language": "C#",
     "type": "item",
     "wts.type": "composition",
     "wts.platform" : "Uwp",
     "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & identity == wts.Feat.DeepLinking & $projectType == TabbedPivot"
-  },
+    "wts.compositionFilter": "$framework == CaliburnMicro & $projectType == Blank & wts.type == project"
+ },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
-  "PrimaryOutputs": [
-  ],
   "symbols": {
     "wts.rootNamespace": {
       "type": "parameter",
