--- templates/Uwp/_composition/CaliburnMicro/Project.BlankTabbedPivot.AddSystemBackButton/.template.config/template.json
@@ -1,31 +1,0 @@
-﻿{
-  "author": "Microsoft",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CaliburnMicro.Project.BlankTabbedPivot.AddSystemBackButton",
-  "tags": {
-    "language": "C#",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & $projectType == Blank|TabbedPivot & wts.type == project"
- },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    },
-    "wts.homePageName": {
-      "type":"parameter",
-      "replaces":"Param_HomeName"
-    }
-  }
-}
