--- templates/Uwp/_composition/CaliburnMicro/Project.SplitView.Page.Settings/.template.config/template.json
@@ -1,29 +1,0 @@
-﻿{
-  "author": "Microsoft",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CaliburnMicro.Project.SplitView.Page.Settings",
-  "tags": {
-    "language": "C#",
-    "type": "item",
-    "wts.type": "composition",
-	"wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$projectType == SplitView & identity == wts.Page.Settings.CaliburnMicro"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [
-  ],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
