+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot.Page.Settings/.template.config/template.json
@@ -1,16 +1,16 @@
 ﻿{
-  "author": "Microsoft Community",
+  "author": "Microsoft",
   "classifications": [
     "Universal"
   ],
-  "name": "CaliburnMicro.DeepLinking.TabbedPivot",
+  "name": "CaliburnMicro.Project.SplitView.TabbedPivot.Page.Settings",
   "tags": {
     "language": "C#",
     "type": "item",
     "wts.type": "composition",
-    "wts.platform" : "Uwp",
+	"wts.platform" : "Uwp",
     "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & identity == wts.Feat.DeepLinking & $projectType == TabbedPivot"
+    "wts.compositionFilter": "$projectType == SplitView|TabbedPivot & identity == wts.Page.Settings.CaliburnMicro"
   },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
@@ -24,10 +24,6 @@
     "wts.itemNamespace": {
       "type": "parameter",
       "replaces": "Param_ItemNamespace"
-    },
-    "wts.homePageName": {
-      "type":"parameter",
-      "replaces":"Param_HomeName"
     }
   }
 }
