+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot.Page.Settings/ViewModels/ShellViewModel_postaction.cs
