+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot.Page.Settings/Views/ShellPage_searchreplace.xaml
