+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot/.template.config/template.json
@@ -3,19 +3,35 @@
   "classifications": [
     "Universal"
   ],
-  "name": "CaliburnMicro.Project.SplitView",
+  "name": "CaliburnMicro.Project.SplitView.TabbedPivot",
   "tags": {
     "language": "C#",
     "type": "item",
     "wts.type": "composition",
 	"wts.platform" : "Uwp",
     "wts.version": "1.0.0",
-    "wts.compositionOrder": "2",
-    "wts.compositionFilter": "$framework == CaliburnMicro & $projectType == SplitView & wts.type == project"
+    "wts.compositionOrder": "1",
+    "wts.compositionFilter": "$framework == CaliburnMicro & $projectType == SplitView|TabbedPivot & wts.type == project",
+    "wts.licenses": "[Windows UI Library](https://aka.ms/winui_license)"
   },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
   "PrimaryOutputs": [
+    {
+      "path": ".\\Helpers\\NavHelper.cs"
+    },
+    {
+      "path": ".\\ViewModels\\ShellViewModel.cs"
+    },
+    {
+      "path": ".\\Views\\ShellPage.xaml"
+    },
+    {
+      "path": ".\\Views\\ShellPage.xaml.cs"
+    },
+    {
+      "path": ".\\Views\\IShellView.cs"
+    }
   ],
   "symbols": {
     "wts.rootNamespace": {
