+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot/App_postaction.xaml
