+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot/Helpers/NavHelper.cs
