+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot/ViewModels/ShellViewModel.cs
