+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot/Views/IShellView.cs
