+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView.TabbedPivot/_postaction.csproj
