--- templates/Uwp/_composition/CaliburnMicro/Project.SplitView/Views/ShellPage.xaml
@@ -1,51 +1,0 @@
-﻿<Page
-    x:Class="wts.ItemName.Views.ShellPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"    
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:i="using:Microsoft.Xaml.Interactivity"
-    xmlns:winui="using:Microsoft.UI.Xaml.Controls"
-    xmlns:behaviors="using:wts.ItemName.Behaviors"
-    xmlns:helpers="using:wts.ItemName.Helpers"
-    xmlns:views="using:wts.ItemName.Views"
-    xmlns:cm="using:Caliburn.Micro"
-    mc:Ignorable="d">
-
-    <winui:NavigationView
-        x:Name="navigationView"
-        IsBackButtonVisible="Visible"
-        IsBackEnabled="{x:Bind ViewModel.IsBackEnabled, Mode=OneWay}"
-        SelectedItem="{x:Bind ViewModel.Selected, Mode=OneWay}"
-        IsSettingsVisible="False"
-        cm:Message.Attach="[Event ItemInvoked] = [Action OnItemInvoked($eventArgs)]"
-        Background="{ThemeResource SystemControlBackgroundAltHighBrush}">
-        <winui:NavigationView.MenuItems>
-            <!--
-            TODO WTS: Change the symbols for each item as appropriate for your app
-            More on Segoe UI Symbol icons: https://docs.microsoft.com/windows/uwp/style/segoe-ui-symbol-font
-            Or to use an IconElement instead of a Symbol see https://github.com/Microsoft/WindowsTemplateStudio/blob/master/docs/projectTypes/navigationpane.md
-            Edit String/en-US/Resources.resw: Add a menu item title for each page
-            -->
-        </winui:NavigationView.MenuItems>        
-        <i:Interaction.Behaviors>
-            <behaviors:NavigationViewHeaderBehavior
-                x:Name="navigationViewHeaderBehavior"
-                DefaultHeader="{x:Bind ViewModel.Selected.Content, Mode=OneWay}">
-                <behaviors:NavigationViewHeaderBehavior.DefaultHeaderTemplate>
-                    <DataTemplate>
-                        <Grid>
-                            <TextBlock
-                                Text="{Binding}"
-                                Style="{ThemeResource TitleTextBlockStyle}"
-                                Margin="{StaticResource SmallLeftRightMargin}" />
-                        </Grid>
-                    </DataTemplate>
-                </behaviors:NavigationViewHeaderBehavior.DefaultHeaderTemplate>
-            </behaviors:NavigationViewHeaderBehavior>
-        </i:Interaction.Behaviors>
-        <Grid>
-            <Frame x:Name="shellFrame" DataContext="{x:Null}" />
-        </Grid>
-    </winui:NavigationView>
-</Page>
