+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView/Views/ShellPage_postaction.xaml
@@ -1,0 +1,29 @@
+﻿<Page
+<!--{[{-->
+    xmlns:i="using:Microsoft.Xaml.Interactivity"
+    xmlns:behaviors="using:wts.ItemName.Behaviors"
+<!--}]}-->
+    mc:Ignorable="d">
+
+    <winui:NavigationView
+        Background="{ThemeResource SystemControlBackgroundAltHighBrush}">
+<!--{[{-->
+        <i:Interaction.Behaviors>
+            <behaviors:NavigationViewHeaderBehavior
+                x:Name="navigationViewHeaderBehavior"
+                DefaultHeader="{x:Bind ViewModel.Selected.Content, Mode=OneWay}">
+                <behaviors:NavigationViewHeaderBehavior.DefaultHeaderTemplate>
+                    <DataTemplate>
+                        <Grid>
+                            <TextBlock
+                                Text="{Binding}"
+                                Style="{ThemeResource TitleTextBlockStyle}"
+                                Margin="{StaticResource SmallLeftRightMargin}" />
+                        </Grid>
+                    </DataTemplate>
+                </behaviors:NavigationViewHeaderBehavior.DefaultHeaderTemplate>
+            </behaviors:NavigationViewHeaderBehavior>
+        </i:Interaction.Behaviors>
+<!--}]}-->
+    </winui:NavigationView>
+</Page>
