+++ templates/Uwp/_composition/CaliburnMicro/Project.SplitView/Views/ShellPage_postaction.xaml.cs
@@ -1,0 +1,14 @@
+﻿namespace wts.ItemName.Views
+{
+    public sealed partial class ShellPage : IShellView
+    {
+        public INavigationService CreateNavigationService(WinRTContainer container)
+        {
+            var navigationService = container.RegisterNavigationService(shellFrame);
+//{[{
+            navigationViewHeaderBehavior.Initialize(navigationService);
+//}]}
+            return navigationService;
+        }
+    }
+}
