+++ templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot.AddNavigationViewItem/.template.config/template.json
@@ -1,16 +1,16 @@
 ﻿{
-  "author": "Microsoft Community",
+  "author": "Nigel Sampson",
   "classifications": [
     "Universal"
   ],
-  "name": "CaliburnMicro.DeepLinking.TabbedPivot",
+  "name": "CaliburnMicro.Project.TabbedPivot.AddNavigationViewItem",
   "tags": {
     "language": "C#",
     "type": "item",
     "wts.type": "composition",
-    "wts.platform" : "Uwp",
+	"wts.platform" : "Uwp",
     "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & identity == wts.Feat.DeepLinking & $projectType == TabbedPivot"
+    "wts.compositionFilter": "wts.type == page & identity != wts.Page.Settings.CaliburnMicro & $framework == CaliburnMicro & $projectType == TabbedPivot"
   },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
@@ -24,10 +24,6 @@
     "wts.itemNamespace": {
       "type": "parameter",
       "replaces": "Param_ItemNamespace"
-    },
-    "wts.homePageName": {
-      "type":"parameter",
-      "replaces":"Param_HomeName"
     }
   }
 }
