+++ templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot.AddNavigationViewItem/Views/ShellPage_postaction.xaml
@@ -1,0 +1,18 @@
+﻿<!--{**-->
+<!--This code adds the new page to the NavigationView control in ShellPage -->
+<!--**}-->
+
+<Page
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <winui:NavigationView
+        Background="{ThemeResource SystemControlBackgroundAltHighBrush}">
+        <winui:NavigationView.MenuItems>
+            <!--^^-->
+            <!--{[{-->
+            <winui:NavigationViewItem x:Uid="Shell_wts.ItemName" helpers:NavHelper.NavigateTo="views:wts.ItemNamePage" />
+            <!--}]}-->
+        </winui:NavigationView.MenuItems>        
+    </winui:NavigationView>
+</Page>
