--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot.AddTabbedPivotItems/.template.config/template.json
@@ -1,28 +1,0 @@
-﻿{
-  "author": "Nigel Sampson",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CaliburnMicro.Project.TabbedPivot.AddTabbedPivotItems",
-  "tags": {
-    "language": "C#",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & $projectType == TabbedPivot & wts.type==page & identity != wts.Page.ImageGallery.CaliburnMicro"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
