--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot.AddTabbedPivotItems/Strings/en-us/Resources_postaction.resw
@@ -1,8 +1,0 @@
-﻿<root>
-  <!--^^-->
-  <!--{[{-->
-  <data name="PivotItem_wts.ItemName.Header" xml:space="preserve">
-    <value>wts.ItemName</value>
-  </data>
-  <!--}]}-->
-</root>
