--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot.AddTabbedPivotItems/ViewModels/PivotViewModel_postaction.cs
@@ -1,13 +1,0 @@
-﻿namespace Param_ItemNamespace.ViewModels
-{
-    public class PivotViewModel : Conductor<Screen>.Collection.OneActive
-    {
-        protected override void OnInitialize()
-        {
-            //^^
-            //{[{
-            Items.Add(new wts.ItemNameViewModel { DisplayName = "PivotItem_wts.ItemName/Header".GetLocalized() });
-            //}]}
-        }
-    }
-}
