--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot/.template.config/template.json
@@ -1,44 +1,0 @@
-﻿{
-  "author": "Nigel Sampson",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CaliburnMicro.Project.TabbedPivot",
-  "tags": {
-    "language": "C#",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == CaliburnMicro & $projectType == TabbedPivot & wts.type == project"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [
-    {
-      "path": ".\\Behaviors\\PivotBehavior.cs"
-    },
-    {
-      "path": ".\\Helpers\\IPivotPage.cs"
-    },
-    {
-      "path": ".\\ViewModels\\PivotViewModel.cs"
-    },
-    {
-      "path": ".\\Views\\PivotPage.xaml"
-    },
-    {
-      "path": ".\\Views\\PivotPage.xaml.cs"
-    }
-  ],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
