--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot/App_postaction.xaml.cs
@@ -1,21 +1,0 @@
-﻿namespace Param_RootNamespace
-{
-    public sealed partial class App
-    {
-        protected override void Configure()
-        {
-            //^^
-            //{[{       
-            _container.PerRequest<PivotViewModel>();
-            //}]}
-        }
-//^^
-//{[{
-
-        private ActivationService CreateActivationService()
-        {
-            return new ActivationService(_container, typeof(PivotViewModel));
-        }
-//}]}
-    }
-}
