--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot/Helpers/IPivotPage.cs
@@ -1,11 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace wts.ItemName.Helpers
-{
-    public interface IPivotPage
-    {
-        Task OnPivotSelectedAsync();
-
-        Task OnPivotUnselectedAsync();
-    }
-}
