--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot/ViewModels/PivotViewModel.cs
@@ -1,12 +1,0 @@
-﻿using wts.ItemName.Helpers;
-using Caliburn.Micro;
-
-namespace wts.ItemName.ViewModels
-{
-    public class PivotViewModel : Conductor<Screen>.Collection.OneActive
-    {
-        protected override void OnInitialize()
-        {
-        }
-    }
-}
