--- templates/Uwp/_composition/CaliburnMicro/Project.TabbedPivot/Views/PivotPage.xaml
@@ -1,35 +1,0 @@
-﻿<Page
-    x:Class="wts.ItemName.Views.PivotPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    Style="{StaticResource PageStyle}"
-    xmlns:behaviors="using:wts.ItemName.Behaviors"
-    xmlns:i="using:Microsoft.Xaml.Interactivity"
-    xmlns:views="using:wts.ItemName.Views"
-    xmlns:cm="using:Caliburn.Micro"
-    mc:Ignorable="d">
-
-    <Grid>
-        <Pivot x:Uid="PivotPage" x:Name="Items">
-            <i:Interaction.Behaviors>
-                <behaviors:PivotBehavior />
-            </i:Interaction.Behaviors>
-             <Pivot.HeaderTemplate>
-                <DataTemplate>
-                    <TextBlock Text="{Binding DisplayName}" />
-                </DataTemplate>
-            </Pivot.HeaderTemplate>
-            <Pivot.ItemTemplate>
-                <DataTemplate>
-                    <ContentControl cm:View.Model="{Binding}"
-                                    HorizontalAlignment="Stretch"
-                                    VerticalAlignment="Stretch"
-                                    HorizontalContentAlignment="Stretch"
-                                    VerticalContentAlignment="Stretch" />
-                </DataTemplate>
-            </Pivot.ItemTemplate>
-        </Pivot>
-    </Grid>
-</Page>
