+++ templates/Uwp/_composition/CaliburnMicro/Project/.template.config/template.json
@@ -9,7 +9,8 @@
     "type": "item",
     "wts.type": "composition",    
     "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",    
+    "wts.version": "1.0.0",
+    "wts.compositionOrder": "0",
     "wts.licenses": "[Caliburn.Micro](https://github.com/Caliburn-Micro/Caliburn.Micro/blob/master/License.txt)",
     "wts.compositionFilter": "$framework == CaliburnMicro & wts.type == project"
   },
